test: four expectations in ./test/ recorded the behaviour this arc changed
The top-level `test/` package is a fifth door neither this arc's gates nor the
review ran, and four of its cells asserted the pre-arc answers. Each is
measured against PostgreSQL 17.11 below; none is a fix to the engine.

**1. `TestDateCastLeavesNonTemporalArithmeticAlone`** wanted
`CAST('1996' AS INT) - 1` as a float64.

    SELECT pg_typeof(CAST('1996' AS INT) - 1)   integer
    SELECT CAST('1996' AS INT) - 1              1995

An integer cast is an integer operand, so the arithmetic over it is integer.
The cell recorded `expr.operandIsInt` not looking through a CAST, which is the
same blind spot that made `ABS(i) * <int8 max>` answer MinInt64 (#849). The
DOMAIN moved; the int4/int8 width is this engine's standing divergence and did
not.

**2. `TestDateTruncOverColumns`** wanted `1996-03-01T00:00:00Z`.

    SELECT date_trunc('year', TIMESTAMP '1996-03-13 14:25:36')::text
      -->  1996-01-01 00:00:00

date_trunc renders the instant the one way this engine renders one, which is
the text the TIMESTAMP column it read produces (#544's second pass). `ld` is
untouched: `last_day_of_month` answers a calendar DATE and always rendered as
one.

**3. `TestShapeOnlyDecodeValuesAreCorrect`** wanted `SUM(LENGTH(url))` = 9 and
4. It is BYTES that it wanted, and the fixture is deliberately not ASCII, so
the two readings are distinguishable:

    group 1   'abc' + '日本'   3 + 2 = 5 characters    3 + 6 = 9 bytes
    group 2   ''    + '😀'     0 + 1 = 1 character     0 + 4 = 4 bytes

    PostgreSQL 17.11:
      length('abc')+length('日本')              5
      octet_length('abc')+octet_length('日本')  9
      length('')+length('😀')                   1
      octet_length('')+octet_length('😀')       4

So 5 and 1 are the server's answers and the shape-only decode is NOT feeding
LENGTH a wrong count — the reviewer's second possibility. What the change does
mean for THIS test is that LENGTH left `logical.shapeLenFuncs` with #856, since
a character count has to read the bytes; the lengths-only path is still what
OCTET_LENGTH takes, the sibling queries above the cell still exercise it, and
what the cell now asserts is that turning that path ON does not change LENGTH's
answer.

**4. `TestMetadataMinMax`** wanted `wantFire: false` for `MIN(i64 + 1)`.

The counter is `physical.MetadataMinMaxPlanned`, and it asks "was this query
ANSWERED from statistics" — it counts the path being PLANNED, and the lift's
own bound reads the manifest in the LOGICAL planner without touching it. So
this is the first of the reviewer's two readings, and the answer is that the
query genuinely is answered from statistics again, which is what the cell's own
text predicted:

> The recovery is real and belongs with #850's: this path HAS the column's min
> and max by construction, so it could fold `agg(col op const)` and reproduce
> the per-row refusal exactly by running the checked kernel on the two EXTREMES.

#850's typed pass does exactly that, one layer up: it runs after annotation,
reads the column's declared type and its statistics, and lifts
`agg(col op const)` only when the extremes prove the per-row form cannot refuse
(|max| + |k| inside the carrier). The aggregate's argument is a bare column by
the time the physical planner looks, so the manifest answers it, as it did
before #841 declined the lift. `wantFire` is true and the rows are unchanged —
and the cell's second arm, which reruns every query with the metadata switch
OFF and compares, is what makes that safe rather than asserted.

`go test -short ./test/` is green; `./internal/engine/expr/` unchanged.

Refs #849, #544, #856, #850

+++ test/date_cast_test.go
@@ -362,7 +362,14 @@ func TestDateCastLeavesNonTemporalArithmeticAlone(t *testing.T) {
 		sql  string
 		want any
 	}{
-		{"SELECT CAST('1996' AS INT) - 1 AS x", float64(1995)},
+		// An INTEGER cast is an integer operand, so the arithmetic over it is
+		// integer: `SELECT pg_typeof(CAST('1996' AS INT) - 1)` is `integer` on
+		// PostgreSQL 17.11 and the value is 1995, measured. This cell wanted a
+		// float64 because `expr.operandIsInt` did not look through a CAST —
+		// the same blind spot that made `ABS(i) * <int8 max>` answer MinInt64
+		// (#849). The DOMAIN is what moved; the width is still this engine's
+		// standing int4/int8 divergence.
+		{"SELECT CAST('1996' AS INT) - 1 AS x", int64(1995)},
 		{"SELECT CAST('12.5' AS DOUBLE) + 1 AS x", 13.5},
 		{"SELECT CAST(1996 AS VARCHAR) AS x", "1996"},
 		// A time-of-day literal has no column type to become, so it keeps
